Renames report-is-editable.guards -> .guard
