+++ front-end/src/app/shared/guards/report-is-editable.guard.ts
